pyproject.toml
[tool.poetry]
name = "arbeitszeugnis"
version = "0.1.0"
description = ""
authors = ["baniasbaabe <banias@hotmail.de>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.9.7 || >3.9.7,<4.0"
litellm = "^1.26.13"
pytest = "^8.0.1"
google-generativeai = "^0.3.2"
python-dotenv = "^1.0.1"
pymupdf = "^1.23.25"
streamlit = "^1.31.1"
spacy = "^3.7.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
